switch to enums, fix unit tests
