+++ services/emailer/pyproject.toml
@@ -16,6 +16,7 @@ sentry-sdk = "^1.27.1"
 launchdarkly-server-sdk = "^8.1.4"
 simple_cloudevent = { git = "https://github.com/daxiom/simple-cloudevent.py" }
 gcp_queue = { git = "https://github.com/bolyachevets/namex.git", subdirectory = "services/pubsub",  branch = "feature-queue-upgrade-python3.12" }
+sbc-common-components = { git = "https://github.com/bcgov/sbc-common-components.git", subdirectory = "python" }
 requests = "^2.31.0"
 jinja2 = "^3.1.2"
 werkzeug = "^3.0.0"
